% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_by_type}
\alias{filter_by_type}
\title{Filter Incidents by Type}
\usage{
filter_by_type(df, type)
}
\arguments{
\item{df}{A dataframe of incidents.}

\item{type}{A string indicating the type of incident (e.g., "malware").}
}
\value{
A filtered dataframe.
}
\description{
Filters a dataframe of cybersecurity incidents by the specified type.
}
