% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_attack_trends}
\alias{plot_attack_trends}
\title{Plot Cyber Attack Trends}
\usage{
plot_attack_trends(df)
}
\arguments{
\item{df}{A dataframe containing a `date` column.}
}
\value{
A ggplot object.
}
\description{
Generates a line plot showing the number of incidents over time.
}
