% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_incidents}
\alias{summarize_incidents}
\title{Summarize Incident Data}
\usage{
summarize_incidents(df)
}
\arguments{
\item{df}{A dataframe of incidents.}
}
\value{
A summary dataframe.
}
\description{
Provides summary statistics of the cybersecurity incident data.
}
